% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_oda.R
\name{add_oda}
\alias{add_oda}
\title{Populate a \code{{peacesciencer}} constructed data set with ODA data}
\usage{
add_oda(data, type = "donor")
}
\arguments{
\item{data}{A data frame created by one of the \verb{create_*()} functions in the \code{{peacesciencer}} R package.}

\item{type}{A character string indicating either \code{"donor"} or \code{"recipient"}. Relevant only if the data object is a state-year level data frame. The value \code{"donor"} (default) means the data will be populated with ODA values given by donor countries. The value \code{"recipient"} means the data will be populated with ODA received by countries.}
}
\value{
\code{add_oda()} returns either a state-year or dyad-year data frame populated with bilateral official development assistance (ODA) commitments and disbursements in millions of 2023 US Dollars. All zero values are true zeros and all \code{NA} values are truly invalid. ODA data runs from 1960 to 2023 for disbursements and from 1966 to 2023 for commitments. If the supplied data object is a dyad-year data frame, ODA values are totals for a donor-recipient pair from state 1 to state 2. If the supplied data is a state-year data frame, ODA values are total bilateral ODA for either a donor (if \code{type = "donor"}) or a recipient (if \code{type = "recipient"}).
}
\description{
The \code{add_oda()} function works by reading in ODA data from DAC2a and DAC3a data tables via a developer API from the OECD data explorer (https://data-explorer.oecd.org/). This approach has certain trade-offs. The main benefit is that the user does not have to eat up extra RAM storing a large-ish dyadic donor-recipient dataset in memory permanently and locally on your own machine. The downside is that to access the data in an active R session you need an internet connection, and calling the data via the OECD's API may take a few seconds to a minute depending on the quality of your connection. I think these costs are well worth the benefit of saving RAM.
}
\details{
This function takes either a dyad-year or state-year dataset made
with the \code{{peacesciencer}} R package and populates it with official
development assistance (ODA) comments and disbursements from the
Organization for Economic Cooperation and Development (OECD), specifically the data tables
DAC2a (disbursements) and DAC3a (commitments). Values are in millions of
2023 U.S. dollars.
}
\examples{
library(tidyverse)
library(peacesciencer)
library(odadata)

# make a dyad-year dataset:
dt <- create_dyadyears(subset_years = 1960:2023) |>
  add_oda() |>
  drop_na(oda_committed) # drops invalid donor-recipient pairs

# make a state-year dataset for aid repients:
dt <- create_stateyears(subset_years = 1960:2023) |>
  add_oda(type = "recipient") |>
  drop_na(oda_committed) # drops non-recipients 
}
